+++ packages/auto-scroll/examples/axis-locking.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import { Table } from './pieces/table';
+
+export default function AxisLocking() {
+  return <Table />;
+}