+++ packages/auto-scroll/src/shared/configuration.ts
@@ -52,5 +52,6 @@ export function getInternalConfig(
     // only allowing limited control over the config at this stage
     maxPixelScrollPerSecond:
       maxPixelScrollPerSecond[provided?.maxScrollSpeed ?? 'standard'],
+    allowedAxis: provided?.allowedAxis ?? 'all',
   };
 }