+++ packages/auto-scroll/src/shared/edges.ts
@@ -1,3 +1,10 @@
-import { Edge } from '../internal-types';
+import { Axis, Edge } from '../internal-types';
 
 export const edges: Edge[] = ['top', 'right', 'bottom', 'left'];
+
+export const edgeAxisLookup: Record<Edge, Axis> = {
+  top: 'vertical',
+  right: 'horizontal',
+  bottom: 'vertical',
+  left: 'horizontal',
+};