+++ packages/auto-scroll/src/shared/is-axis-allowed.ts
@@ -1,0 +1,5 @@
+import { AllowedAxis, Axis } from '../internal-types';
+
+export function isAxisAllowed(axis: Axis, allowedAxis: AllowedAxis) {
+  return allowedAxis === 'all' || axis === allowedAxis;
+}