+++ packages/auto-scroll/src/unsafe-overflow/get-scroll-by.ts
@@ -6,16 +6,17 @@ import type {
 
 import type { Axis, Edge, InternalConfig, Spacing } from '../internal-types';
 import { canScrollOnEdge } from '../shared/can-scroll-on-edge';
-import { edges } from '../shared/edges';
+import { edgeAxisLookup, edges } from '../shared/edges';
 import { markAndGetEngagement } from '../shared/engagement-history';
 import { getScrollChange } from '../shared/get-scroll-change';
+import { isAxisAllowed } from '../shared/is-axis-allowed';
 import { isWithin } from '../shared/is-within';
 
 import { getHitbox } from './hitbox';
 import {
-  HitboxSpacing,
-  ProvidedHitboxSpacing,
-  UnsafeOverflowAutoScrollArgs,
+  type HitboxSpacing,
+  type ProvidedHitboxSpacing,
+  type UnsafeOverflowAutoScrollArgs,
 } from './types';
 
 export type HitboxForEdge = {
@@ -138,8 +139,16 @@ export function getScrollBy<DragType extends AllDragTypes>({
   // as being engaged with to start applying time dampening
   const engagement = markAndGetEngagement(entry.element);
 
-  const scrollableEdges: HitboxForEdge[] = inHitboxForEdge.filter(value =>
-    canScrollOnEdge[value.edge](entry.element),
+  const { allowedAxis } = config;
+
+  // Note: changing the allowed axis during a drag will not
+  // reset time dampening. It was decided it would be too
+  // complex to implement initially, and we can add it
+  // later if needed.
+  const scrollableEdges: HitboxForEdge[] = inHitboxForEdge.filter(
+    value =>
+      isAxisAllowed(edgeAxisLookup[value.edge], allowedAxis) &&
+      canScrollOnEdge[value.edge](entry.element),
   );
 
   // Nothing can be scrolled